Rimossi commenti e sostituiti componenti non utilizzati
