+++ public/app/features/dashboard/dashgrid/PanelStateWrapper.tsx
@@ -19,8 +19,9 @@ import {
   TimeRange,
   toDataFrameDTO,
   toUtc,
+  TypedVariableModel,
 } from '@grafana/data';
-import { RefreshEvent } from '@grafana/runtime';
+import { RefreshEvent, getTemplateSrv } from '@grafana/runtime';
 import { VizLegendOptions } from '@grafana/schema';
 import {
   ErrorBoundary,
@@ -99,14 +100,14 @@ export class PanelStateWrapper extends PureComponent<Props, State> {
         sync: this.getSync,
         onSeriesColorChange: this.onSeriesColorChange,
         onToggleSeriesVisibility: this.onSeriesVisibilityChange,
-        onAnnotationCreate: this.onAnnotationCreate,
-        onAnnotationUpdate: this.onAnnotationUpdate,
-        onAnnotationDelete: this.onAnnotationDelete,
+        onAnnotationCreate: undefined,
+        onAnnotationUpdate: undefined,
+        onAnnotationDelete: undefined, // Annotation delete 
         onInstanceStateChange: this.onInstanceStateChange,
         onToggleLegendSort: this.onToggleLegendSort,
-        canAddAnnotations: props.dashboard.canAddAnnotations.bind(props.dashboard),
-        canEditAnnotations: props.dashboard.canEditAnnotations.bind(props.dashboard),
-        canDeleteAnnotations: props.dashboard.canDeleteAnnotations.bind(props.dashboard),
+        canAddAnnotations: undefined,
+        canEditAnnotations: undefined,
+        canDeleteAnnotations: undefined,// Annotation delete 
         onAddAdHocFilter: this.onAddAdHocFilter,
         onUpdateData: this.onUpdateData,
       },
@@ -190,14 +191,14 @@ export class PanelStateWrapper extends PureComponent<Props, State> {
     };
   }
 
-  componentDidMount() {
-    const { panel, dashboard } = this.props;
+  initializePanel() {
+    const { dashboard, panel } = this.props;
 
     // Subscribe to panel events
     this.subs.add(panel.events.subscribe(RefreshEvent, this.onRefresh));
     this.subs.add(panel.events.subscribe(RenderEvent, this.onRender));
 
-    dashboard.panelInitialized(this.props.panel);
+    dashboard.panelInitialized(panel);
 
     // Move snapshot data into the query response
     if (this.hasPanelSnapshot) {
@@ -225,6 +226,43 @@ export class PanelStateWrapper extends PureComponent<Props, State> {
     liveTimer.listen(this);
   }
 
+  componentDidMount() {
+    this.initializePanel();
+
+    const receiveMessage = (event: any) => {
+      if (event.data.variables != undefined) {
+        const change = event.data.variables;
+        const srv = getTemplateSrv();
+        const variables = srv.getVariables();
+        const newVariables: TypedVariableModel[] = [];
+        let tmp: TypedVariableModel;
+        let newV: TypedVariableModel;
+        change.forEach((c) => {
+          tmp = variables.find((v) => v.name === c.key);
+          if (tmp != undefined) {
+            newV = { ...tmp };
+            newV.current = { ...tmp.current, value: c.value };
+            newVariables.push(newV);
+          }
+        });
+        if (newVariables.length > 0) {
+          srv.init(newVariables);
+          this.onRefresh();
+        }
+      }
+      if (event.data.timeRange != undefined) {
+        this.onChangeTimeRange(event.data.timeRange);
+      }
+    };
+
+    window.addEventListener('message', receiveMessage);
+
+    return () => {
+      window.removeEventListener('message', receiveMessage);
+    };
+
+  }
+
   componentWillUnmount() {
     this.subs.unsubscribe();
     liveTimer.remove(this);
@@ -243,8 +281,8 @@ export class PanelStateWrapper extends PureComponent<Props, State> {
     this.setState({ liveTime });
   }
 
-  componentDidUpdate(prevProps: Props) {
-    const { isInView, width } = this.props;
+  componentDidUpdate(prevProps: Props, prevState: State) {
+    const { isInView, width, panel } = this.props;
     const { context } = this.state;
 
     const app = this.getPanelContextApp();
@@ -272,13 +310,14 @@ export class PanelStateWrapper extends PureComponent<Props, State> {
     if (width !== prevProps.width) {
       liveTimer.updateInterval(this);
     }
+
   }
 
   // Updates the response with information from the stream
   // The next is outside a react synthetic event so setState is not batched
   // So in this context we can only do a single call to setState
   onDataUpdate(data: PanelData) {
-    const { dashboard, panel, plugin } = this.props;
+    const { dashboard, plugin, panel } = this.props;
 
     // Ignore this data update if we are now a non data panel
     if (plugin.meta.skipDataQuery) {
@@ -326,15 +365,13 @@ export class PanelStateWrapper extends PureComponent<Props, State> {
   }
 
   onRefresh = () => {
-    const { dashboard, panel, isInView, width } = this.props;
-
+    const { dashboard, isInView, width, panel } = this.props;
     if (!isInView) {
       this.setState({ refreshWhenInView: true });
       return;
     }
 
     const timeData = applyPanelTimeOverrides(panel, this.timeSrv.timeRange());
-
     // Issue Query
     if (this.wantsQueryExecution) {
       if (width < 0) {
@@ -440,6 +477,10 @@ export class PanelStateWrapper extends PureComponent<Props, State> {
     });
   };
 
+  onChangeTimeRangeOverride = (timeRange: AbsoluteTimeRange) => {
+    return;
+  };
+
   shouldSignalRenderingCompleted(loadingState: LoadingState, pluginMeta: PanelPluginMeta) {
     return loadingState === LoadingState.Done || loadingState === LoadingState.Error || pluginMeta.skipDataQuery;
   }
@@ -468,7 +509,7 @@ export class PanelStateWrapper extends PureComponent<Props, State> {
   };
 
   renderPanelContent(innerWidth: number, innerHeight: number) {
-    const { panel, plugin, dashboard } = this.props;
+    const { plugin, dashboard, panel } = this.props;
     const { renderCounter, data } = this.state;
     const { state: loadingState } = data;
 
@@ -490,7 +531,6 @@ export class PanelStateWrapper extends PureComponent<Props, State> {
     // Update the event filter (dashboard settings may have changed)
     // Yes this is called ever render for a function that is triggered on every mouse move
     this.eventFilter.onlyLocal = dashboard.graphTooltip === 0;
-
     return (
       <>
         <PanelContextProvider value={this.state.context}>
@@ -509,7 +549,7 @@ export class PanelStateWrapper extends PureComponent<Props, State> {
             replaceVariables={panel.replaceVariables}
             onOptionsChange={this.onOptionsChange}
             onFieldConfigChange={this.onFieldConfigChange}
-            onChangeTimeRange={this.onChangeTimeRange}
+            onChangeTimeRange={this.onChangeTimeRangeOverride}
             eventBus={dashboard.events}
           />
         </PanelContextProvider>
@@ -518,11 +558,11 @@ export class PanelStateWrapper extends PureComponent<Props, State> {
   }
 
   render() {
-    const { dashboard, panel, width, height, plugin } = this.props;
+    const { dashboard, width, height, plugin, panel } = this.props;
     const { errorMessage, data } = this.state;
     const { transparent } = panel;
 
-    const panelChromeProps = getPanelChromeProps({ ...this.props, data });
+    const panelChromeProps = getPanelChromeProps({ ...this.props, data, panel });
 
     // Shift the hover menu down if it's on the top row so it doesn't get clipped by topnav
     const hoverHeaderOffset = (panel.gridPos?.y ?? 0) === 0 ? -16 : undefined;
